Add container and scripts for report visualization
Add scripts to build, start, and stop a Docker container which
allows the HTML reports, including visualizations, to be viewed
locally without having to setup a web server with PHP support.

Update the existing man pages to reference the new scripts
kivis-build, kivis-start, and kivis-stop, and add new manual
pages for those scripts.

Update the linuxki_install script to automatically build the
docker container when LinuxKI is installed, if Docker is already
installed. If Docker is installed after LinuxKI is installed,
the kivis-build script can be run manually to build the
container image.

Signed-off-by: Christopher Voltz <christopher@voltz.ws>