Tell user if wrong package is installed
Modify kivis-start to notify the user if the wrong package is
installed and exit with an error.

Debian has a package called docker which also provides a docker
command but is not related to Docker containers so an
unsuspecting user may think they have Docker installed when they
don't. The error message provides the URLs to Docker
documentation on how to install Docker in various
distributions.