Remove sudo from kivis-build
Remove sudo from the command to build the linuxki Docker image. This removes
the requirement for the user to have sudo privileges. Docker requires that the
user building the image either be in the docker group or have sudo privileges.
In the latter case, kivis-build must be run manually (using sudo) before
calling kivis-start or the user will get an error.